+++ cpu/port.h
@@ -5,6 +5,6 @@
 
 uint8_t portByteIn(uint16_t);
 uint16_t portWordIn(uint16_t);
-void portByteOut(uint32_t, uint32_t);
+void portByteOut(uint16_t, uint8_t);
 void portWordOut(uint16_t, uint16_t);
 #endif
